[refactor] rename SessionRepository in SessionProjectionRepository
