+++ app/Http/Providers/AppServiceProvider.php
@@ -14,7 +14,7 @@ class AppServiceProvider extends ServiceProvider {
         $this->app->bind('App\Domain\IEventPublisher', 'App\Infrastructure\EventPublisher');
 		$this->app->bind('App\Infrastructure\IEventStore', 'App\Infrastructure\FileEventStore');
         $this->app->bind('App\Infrastructure\IProjectionStore', 'App\Infrastructure\FileProjectionStore');
-        $this->app->bind('App\Domain\Identity\ISessionRepository', 'App\Infrastructure\Identity\SessionRepository');
+        $this->app->bind('App\Domain\Identity\ISessionProjectionRepository', 'App\Infrastructure\Identity\SessionProjectionRepository');
 	}
 
 	/**
